refactor user messages
